Fix bug of checking result before it's finished
The slackbot looks at the judging entries to see whether they are correct or not. 
However, the entry is first created with a result of 'null', where that null result is then replaced with the actual outcome. 
It did not take this into account, leading to missed messages, but now it does take it into account and waits for the judging to complete.